Code review improvement
