+++ lib/shared/widgets/bubble.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:givt_app_kids_web/shared/widgets/bubble_draw_triangle.dart';
 
 class Bubble extends StatelessWidget {
   const Bubble({
@@ -32,7 +33,7 @@ class Bubble extends StatelessWidget {
             ),
             CustomPaint(
               size: Size(triangleSize, triangleSize / 2),
-              painter: DrawTriangle(
+              painter: BubbleDrawTriangle(
                 color: color,
                 direction: direction,
               ),
@@ -44,7 +45,7 @@ class Bubble extends StatelessWidget {
           children: [
             CustomPaint(
               size: Size(triangleSize / 2, triangleSize),
-              painter: DrawTriangle(
+              painter: BubbleDrawTriangle(
                 color: color,
                 direction: direction,
               ),
@@ -60,46 +61,5 @@ class Bubble extends StatelessWidget {
           ],
         );
     }
-  }
-}
-
-enum BubbleDirection {
-  above,
-  atRight,
-}
-
-class DrawTriangle extends CustomPainter {
-  DrawTriangle({
-    required this.color,
-    required this.direction,
-  });
-
-  final Color color;
-  final BubbleDirection direction;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final path = Path();
-    switch (direction) {
-      case BubbleDirection.above:
-        path.moveTo(0, 0);
-        path.lineTo(size.width / 2, size.height);
-        path.lineTo(size.width, 0);
-        path.close();
-        canvas.drawPath(path, Paint()..color = color);
-        break;
-      case BubbleDirection.atRight:
-        path.moveTo(size.width, 0);
-        path.lineTo(0, size.height / 2);
-        path.lineTo(size.width, size.height);
-        path.close();
-        canvas.drawPath(path, Paint()..color = color);
-        break;
-    }
-  }
-
-  @override
-  bool shouldRepaint(CustomPainter oldDelegate) {
-    return false;
   }
 }
